// src/store/types.ts
export interface IState {
  token: string | null;
  pending: boolean;
  error: string | null;
  cart: CartItem[];
  orders: Order[];
  refs: Refs;
  favoriteProducts: Favorite[] | [];
  foundProducts: Product[];
  promoProducts: UserAction[];
  user: User;
  suggestions: {streets: any; houses: any};
  catalog: Category[] | [];
  products: {[key: string]: Product[]};
  reviews: Review[];
  exists: boolean;
  phoneNumber: string;
}